Se arregla bug de parseo de datos double a int

+++ src/Frontend/src/hooks/useInsertReservation.jsx
@@ -33,7 +33,7 @@ const useInsertReservation = (reservation) => {
             Reservation_Date,
             Name_Service: NewServices[index].Name_Service,
             Price: NewServices[index].Price,
-            Quantity: parseInt(NewServices[index].Quantity),
+            Quantity: parseFloat(NewServices[index].Quantity),
             Currency: NewServices[index].Currency
           });
         })
